requirements.txt
streamlit
google-generativeai
PyPDF2
chromadb
sentence-transformers
langchain
